+++ CHANGELOG.md
@@ -2,4 +2,17 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.1.0](https://github.com/lzaaldian/bootstrap-wordpress-custom-theme/compare/v1.0.0...v1.1.0) (2021-05-01)
+
+
+### Features
+
+* add front-page design and initial design using Bootstrap ([b9abb8e](https://github.com/lzaaldian/bootstrap-wordpress-custom-theme/commit/b9abb8e383d02bbefad1fee5065d626e673f98cd))
+* add home page design with sticky footer ([d73eb6c](https://github.com/lzaaldian/bootstrap-wordpress-custom-theme/commit/d73eb6cd9e31a2aac5c4706ea040d17c0900b57b))
+
+
+### Bug Fixes
+
+* padding between navbar and content ([70b8f77](https://github.com/lzaaldian/bootstrap-wordpress-custom-theme/commit/70b8f772af5ec95a4b1c9e4f65b4e79ed1f5616a))
+
 ## 1.0.0 (2021-04-11)